+++ src/components/organisms/SideBar.tsx
@@ -14,13 +14,6 @@ export const SideBar = () => {
             makeKeyNode({ x: 0, y: 0 }, { keySig: init.sig, isMajor: init.isMajor })
         ])
     }
-    const setDraggable = (isDraggable: boolean) => {
-        setNodes(nds =>
-            nds.map(node => {
-                node.draggable = isDraggable
-                return node;
-            }))
-    }
     return (
         <div>
             <div>
@@ -29,12 +22,6 @@ export const SideBar = () => {
                 </button>
                 <button onClick={initialize}>
                     init
-                </button>
-                <button onClick={() => setDraggable(true)}>
-                    draggable
-                </button>
-                <button onClick={() => setDraggable(false)}>
-                    unDraggable
                 </button>
             </div>
             <div>
